Refactored workflow store to for better maintainability
