+++ apps/web/src/app-types/global.d.ts
