+++ apps/web/src/app-types/workflow/edges.ts
@@ -1,0 +1,6 @@
+export type EdgeData = {
+    id: string;
+    workflowId?: string;
+    toNodeId?: string;
+    fromNodeId?: string;
+};
