+++ apps/web/src/components/agents/layouts/base-agent-layout.tsx
@@ -3,9 +3,8 @@ import { CSSProperties } from 'react';
 
 import './agent-card.styles.css';
 import cloudHeadImage from '@public/pink-cloud-head.png';
-import { AgentData } from '@lib/definitions';
-import { AgentState } from '@lib/definitions';
-import { AgentStatus } from '@lib/definitions';
+import { AgentData } from '@app-types/agent';
+import { AgentState } from '@app-types/agent';
 import { InfoIcon } from '@components/icons/info-icon';
 import { RefreshIcon } from '@components/icons/refresh-icon';
 import { TargetIcon } from '@components/icons/target-icon';
