+++ apps/web/src/lib/agents/browser/navigation.ts
