+++ apps/web/src/lib/dagre/dagre.ts
@@ -10,11 +10,9 @@ import { ROOT_NODE_X } from '@config/layout.const';
 import { ROOT_NODE_Y } from '@config/layout.const';
 import { NODE_SPACING_X } from '@config/layout.const';
 import { NODE_SPACING_Y } from '@config/layout.const';
-import { NodeData } from '@lib/definitions';
+import type { NodeData } from '@app-types/workflow';
 
-const {
-    WorkflowNode: { WIDTH, HEIGHT }
-} = Config;
+const { WorkflowNode: { WIDTH, HEIGHT } } = Config;
 
 export const layoutElements = (nodes: Node<NodeData>[], edges: Edge[]) => {
     const isInitialNodes = nodes.every(node => node?.type?.includes('initial'));
